+++ database/migrations/2025_04_19_080534_create_sprints_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('planner_sprints', function (Blueprint $table) {
-            $table->id();
-            $table->uuid('uuid')->unique();
-            $table->foreignId('project_id')->constrained('planner_projects')->onDelete('cascade');
-            $table->string('name');
-            $table->date('start_date')->nullable();
-            $table->date('end_date')->nullable();
-            $table->unsignedInteger('order')->default(0);
-            $table->foreignId('user_id')->nullable()->constrained('users')->nullOnDelete();
-            $table->foreignId('team_id')->nullable()->constrained('teams')->nullOnDelete();
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('planner_sprints')) {
+            Schema::create('planner_sprints', function (Blueprint $table) {
+                $table->id();
+                $table->uuid('uuid')->unique();
+                $table->foreignId('project_id')->constrained('planner_projects')->onDelete('cascade');
+                $table->string('name');
+                $table->date('start_date')->nullable();
+                $table->date('end_date')->nullable();
+                $table->unsignedInteger('order')->default(0);
+                $table->foreignId('user_id')->nullable()->constrained('users')->nullOnDelete();
+                $table->foreignId('team_id')->nullable()->constrained('teams')->nullOnDelete();
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
